Add files via upload
When gamma function is applied to an image, the self_image becomes as reference image, and we usually work in a frame where the reference image is the one extrated from TPS (what it should be). So gamma function should be applied to expected_image, so the reference image is the expected and the comparison image is the actual_image.