image.py updated
For a global gamma analysis, normalize parameter is set to True, but for a local gamma normalize should be set to False. Once it is set to False, and you don't normalize the images, in the numerator of the gamma function you have an absolute difference, then in the denominator there should be also an absolute number, however there isn't. Multiplying the dose difference (doseTA) parameter there must be the expected fluence, so we may apply the local dose difference criterion correctly.

Since the normalization should be done to the same point for both fluence maps and they are self-normalized,